fixup! fixup! fixup! fixup! Refactor tests in GIOS
